requirements.txt
Flask
pillow
pytesseract
PyMuPDF
googletrans
transformers
torch
asyncio
re